pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drun"
version = "0.3.0"
description = "Minimal HTTP API test runner (MVP)"
requires-python = ">=3.10"
authors = [{ name = "Drun Team" }]
readme = "README.md"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.6",
  "jmespath>=1.0",
  "PyYAML>=6.0",
  "rich>=13.7",
  "typer>=0.12",
]

[project.scripts]
drun = "drun.cli:app"

[tool.setuptools.packages.find]
include = ["drun*"]
exclude = ["spec*", "testcases*", "testsuites*"]
